Reader parses values written prior to encoding
With the addition of setting encoding on write (yay!), it is possible
that existing persisted statements were written without encoding. When
attempting to read those previously written statements, the reader
raises an Encoding::UndefinedConversionError.

By catching that error and sending it through the unescape sequence,
it is possible to apply the unescape/encoding correct behavior.

Should the RDF::NTriples::Reader.unescape method be extracted as a
service method on RDF? Given that this pull request is setting up a
case where the Ancestor class is calling a method on a Descendant
class.

This appears to be an issue when moving from 1.0.9 to 1.0.10

https://github.com/ruby-rdf/rdf/compare/1.0.9...1.0.10, in particular
https://github.com/ruby-rdf/rdf/compare/1.0.9...1.0.10#diff-94c44c6c4ab4c6e92f428841d0e827c3L60